+++ p2panda-rs/src/schema/mod.rs
@@ -9,10 +9,8 @@ use cddl::validator::cbor;
 
 mod message;
 mod schema;
-mod schema_v2;
 
 pub use message::MESSAGE_SCHEMA;
-pub use schema::UserSchema;
 
 /// Custom error types of schema validation.
 pub mod error {
